+++ crates/web-client/test/remote_keystore.test.ts
@@ -182,7 +182,7 @@ test.describe("remote keystore", () => {
       );
 
       // This call should trigger the sign callback
-      await client.newTransaction(faucet.id(), txRequest);
+      await client.executeTransaction(faucet.id(), txRequest);
 
       return { faucetPubKey, signPubKey };
     });